+++ packages/tools/file/src/path.spec.js
@@ -2,12 +2,26 @@ import {strict as assert} from 'assert';
 
 import {
 	filterJsonExtensions,
+	getPathExt,
 	hasAnyExtensionOf,
+	isFileWithExt,
 	renameToExtension,
 	resolveToDir,
 } from './path';
 
 describe('path', function() {
+	describe('getPathExt', function() {
+		it('should return the extension of the provided file path', function() {
+			assert.deepStrictEqual(
+				getPathExt('foo/bar.txt'),
+				'txt'
+			);
+			assert.deepStrictEqual(
+				getPathExt('foo/bar.todo.md'),
+				'md'
+			);
+		});
+	});
 	describe('hasAnyExtensionOf', function() {
 		it('should detect if a file name has one of the provided extensions', function() {
 			const isJsonOrGeojson = hasAnyExtensionOf(['.json', '.geojson']);
@@ -28,6 +42,14 @@ describe('path', function() {
 				filterJsonExtensions(['file.json', 'file.geojson', 'file.csv']),
 				['file.json', 'file.geojson']
 			);
+		});
+	});
+	describe('isFileWithExt', function() {
+		it('should return a function that returns true if the input file name has the provided extension', function() {
+			const isJson = isFileWithExt('json');
+
+			assert.deepStrictEqual(isJson('file.json'), true);
+			assert.deepStrictEqual(isJson('file.txt'), false);
 		});
 	});
 	describe('renameToExtension', function() {